Seme fix demo classes flow Fix name bank text field
