+++ ExampleiOS/Source/Classes/UserStrories/Main/UIKit/Cells/BaseCell/TralingElements/TralingElementsViewController.swift
@@ -1,15 +1,15 @@
 //
-//  ImageBaseCellTralingMenuController.swift
+//  TralingComponentsViewController.swift
 //  ExampleiOS
 //
-//  Created on 20.11.2020.
+//  Created on 18.11.2020.
 //
 
 import AdmiralUIKit
 import AdmiralTheme
 import UIKit
 
-final class ImageBaseCellTralingViewController: ScrollViewController {
+final class TralingElementsViewController: ScrollViewController {
     
     var cells: [ListViewCell] = []
     
@@ -48,18 +48,19 @@ final class ImageBaseCellTralingViewController: ScrollViewController {
             $0.isEnabled = isEnabled
         })
     }
-    
+
     private func configureCells() {
         cells.append(configureArrowCell())
         cells.append(configureRadioButtonCell())
         cells.append(configureCheckBoxCell())
         cells.append(configureSwitchListViewCell())
+        cells.append(configureCurcleIconListViewCell())
         cells.append(configureCardListViewCell())
         cells.append(configureIconListViewCell())
         cells.append(configureDatePercentListViewCell())
         cells.append(configureImageWithSubtitleListViewCell())
-        cells.append(configureCurcleIconListViewCell())
-
+        cells.append(configureSubtitleWithImageListViewCell())
+        
         for index in 0..<cells.count {
             let gesture = UITapGestureRecognizer(target: self, action: #selector(tapCell(_:)))
             cells[index].isUserInteractionEnabled = true
@@ -67,114 +68,103 @@ final class ImageBaseCellTralingViewController: ScrollViewController {
             cells[index].tag = index
         }
     }
-    
+
     @objc private func tapCell(_ tap: UITapGestureRecognizer) {
         guard let view = tap.view as? ListViewCell else { return }
-        
         for cell in cells {
-            if cell == view {
-                cell.isHighlighted = true
-            } else {
-                cell.isHighlighted = false
-            }
+            let isSelectedCell = cell == view
+            cell.isHighlighted = isSelectedCell
         }
     }
-    
-    private func configureArrowCell() -> ListCell<LeadingCardListView, TitleListView, ArrowListView> {
+
+    private func configureArrowCell() -> ListCell<ListCellEmpty, TitleListView, ArrowListView> {
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: ArrowListView())
+        return ListCell(centerView: titleListView, tralingView: ArrowListView())
     }
     
-    private func configureRadioButtonCell() -> ListCell<LeadingCardListView, TitleListView, RadioButtonListView> {
+    private func configureRadioButtonCell() -> ListCell<ListCellEmpty, TitleListView, RadioButtonListView> {
         let radioButtonView = RadioButtonListView()
-        radioButtonView.radioButtonAccesibilityId = "RadioButtonListViewButton"
-        radioButtonView.accessibilityIdentifier = "RadioButtonListView"
+        radioButtonView.isSelected = true
+        radioButtonView.didSelect = {}
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: radioButtonView)
+        return ListCell(centerView: titleListView, tralingView: radioButtonView)
     }
     
-    private func configureCheckBoxCell() -> ListCell<LeadingCardListView, TitleListView, CheckBoxListView> {
+    private func configureCheckBoxCell() -> ListCell<ListCellEmpty, TitleListView, CheckBoxListView> {
         let checkBoxCell = CheckBoxListView()
         checkBoxCell.checkBoxAccesibilityId = "CheckBoxListViewCheckBox"
         checkBoxCell.accessibilityIdentifier = "CheckBoxListView"
         checkBoxCell.isSelected = true
+        checkBoxCell.didSelect = {}
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: checkBoxCell)
+        return ListCell(centerView: titleListView, tralingView: checkBoxCell)
     }
     
-    private func configureSwitchListViewCell() -> ListCell<LeadingCardListView, TitleListView, SwitchListView> {
+    private func configureSwitchListViewCell() -> ListCell<ListCellEmpty, TitleListView, SwitchListView> {
         let switchListView = SwitchListView()
         switchListView.didSelect = { isOn in
             print(isOn)
         }
         switchListView.customSwitchAccesibilityId = "SwitchListViewSwitcher"
         switchListView.accessibilityIdentifier = "SwitchListView"
+        switchListView.isControlSelected = true
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: switchListView)
+        return ListCell(centerView: titleListView, tralingView: switchListView)
     }
     
-    private func configureCardListViewCell() -> ListCell<LeadingCardListView, TitleListView, CardListView> {
+    private func configureCardListViewCell() -> ListCell<ListCellEmpty, TitleListView, CardListView> {
         let сardListView = CardListView()
         сardListView.cardImage = Asset.Card.visa.image
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: сardListView)
+        return ListCell(centerView: titleListView, tralingView: сardListView)
     }
     
-    private func configureIconListViewCell() -> ListCell<LeadingCardListView, TitleListView, IconListView> {
+    private func configureIconListViewCell() -> ListCell<ListCellEmpty, TitleListView, IconListView> {
         let iconListView = IconListView()
         iconListView.image = Asset.Main.calendar.image
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: iconListView)
+        return ListCell(centerView: titleListView, tralingView: iconListView)
     }
     
-    private func configureDatePercentListViewCell() -> ListCell<LeadingCardListView, TitleListView, DatePercentListView> {
+    private func configureDatePercentListViewCell() -> ListCell<ListCellEmpty, TitleListView, DatePercentListView> {
         let datePercentListView = DatePercentListView()
         datePercentListView.date = "Date"
         datePercentListView.percent = "Percent"
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: datePercentListView)
+        return ListCell(centerView: titleListView, tralingView: datePercentListView)
     }
     
-    private func configureImageWithSubtitleListViewCell() -> ListCell<LeadingCardListView, TitleListView, ImageWithSubtitleListView> {
+    private func configureImageWithSubtitleListViewCell() -> ListCell<ListCellEmpty, TitleListView, ImageWithSubtitleListView> {
         let imageWithSubtitleListView = ImageWithSubtitleListView()
         imageWithSubtitleListView.subtitle = "Subtitle"
         imageWithSubtitleListView.image = Asset.Card.mir.image
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: imageWithSubtitleListView)
+        return ListCell(centerView: titleListView, tralingView: imageWithSubtitleListView)
     }
     
-    private func configureCurcleIconListViewCell() -> ListCell<LeadingCardListView, TitleListView, CurcleIconListView> {
+    private func configureSubtitleWithImageListViewCell() -> ListCell<ListCellEmpty, TitleListView, SubtitleWithImageListView> {
+        let imageWithSubtitleListView = SubtitleWithImageListView()
+        imageWithSubtitleListView.subtitle = "Subtitle"
+        imageWithSubtitleListView.image = Asset.Card.mir.image
+        let titleListView = TitleListView()
+        titleListView.title = "Title"
+        return ListCell(centerView: titleListView, tralingView: imageWithSubtitleListView)
+    }
+    
+    private func configureCurcleIconListViewCell() -> ListCell<ListCellEmpty, TitleListView, CurcleIconListView> {
         let curcleIconListView = CurcleIconListView()
         curcleIconListView.image = Asset.Card.rnb.image
         let titleListView = TitleListView()
         titleListView.title = "Title"
-        let cardListView = LeadingCardListView()
-        cardListView.cardImage = Asset.Card.visa.image
-        return ListCell(leadingView: cardListView, centerView: titleListView, tralingView: curcleIconListView)
+        return ListCell(centerView: titleListView, tralingView: curcleIconListView)
     }
-
+    
 }
